+++ Program 2 for Programming Systems/node.h
