+++ Program 2 for Programming Systems/table.h
@@ -3,7 +3,7 @@
  *
  */
 
-#include "low.h"
+#include "node.h"
 
 const int NUM = 3;
